Modify configuration file and handling
The configuration file was a mess with the changes introduced in
c88bc98199d1677449c839371bf7558d5d678db5, so it has been completely
rearranged, placing all entries' options into the entry itself. The
handling of the configuration has been drastically rewritten such that
only entries defined in the configuration file are printed, they are
printed in the order given (implements HorlogeSkynet/archey4#57), and
potentially given a switch away from JSON we could have entries defined
multiple times (this could be a good approach for a highly customisable
disk entry).

The configuration handling has been simplified - in my opinion, while
the code to update the configuration through subsequent files provided
override was not prevented was nice, but unnecessary. The default
configuration defined in the `Configuration` class was also unsightly,
so this has been removed, and the logic now takes the configuration file
with highest precedence. A configuration file is now a requirement.

Additionally, renamed the `module` base class to `entry`, since `module`
is easy to confuse with Python modules. An `__init__` method has also
been added to it, defining the `name` attribute of the calling class,
which incidentally fixes the linting errors on this file.

All changes:
* Combined entries and their options in config.json.
* Configuration logic now follows the above pattern (removed the
  recursive updating).
* Removed the binding `get` method from `Configuration`, replaced with
  `__getitem__` to emulate the inbuilt `dict` behaviour (`self[key]`)
* Rewrote `__main__.py` to dynamically import entries (and initialise
  and attach them to the `output` object) based on the configuration.
* Renamed some modules to reflect the class in them to enable the
  changes in `__main__.py`.
* Renamed `module` base class to `entry`.
* Made necessary rewrites in all entries for new configuration.
* Made all entries call the `entry` base class's `__init__` to populate
  their name attribute instead of having each class do it individually.
* Some tweaks to error messages, and grammar.
* Removed a redundant `Pool` import accidentally included last commit,
  and cleaned up some areas with trailing whitespace to fix linting
  errors.