Fix stickytape build
Adds a python script to get the required arguments for stickytape and
run it on the CLI. I'm not entirely sure whether I like this solution,
but it's better than having our users enter a several lines long
command.

* Updated README with new stickytape instructions.
* Updated Travis to use new build instructions.

The new build instructions can be reverted on a new release of
stickytape incorporating the new CLI features. The `master` branch seems
stable enough, however, to use directly like this as a stop-gap
solution.