[DISK] Rewrite disk entry.
The entire entry has been rewritten, and now relies exclusively on a
single call to `df -P -k`.

* Adds support for multi-line disk output (fixes #55).
* Fixes #62 by using a standard `df` call.
* Adds configurability to disk labels, and which disks to show.
  * Modifies `config.json`, `configuration.py` and README accordingly.

More detail:

Our single call to `df` now saves the results in a table-like format,
so we can simply reuse this to extract our disks' mount-points, device
paths and used/total blocks.

This saves us on slow system calls, and the new structure allows for a
more suitable list of dicts (which each represent a filesystem) to be
stored in the `value` attribute of the entry. This in turn provides us
with a highly configurable entry!