[DISK] Use dictionary-focussed approach.
Basically rewrites everything in such a way that we now deal with dicts
whose keys are mount points, and values are dicts of other data for that
disk. Mount points are used as keys since they are globally unique.

This simplifies quite a lot of the internal logic of `Disk`, likely
resulting in faster execution (since it eradicates a few now-unneeded
iterations of `df`'s output).

Additionally:

+ Cleans up consistency in variable names and documentation, to use:
  * `mount_point` and "mount point"
  * `device_path` and "device path"
  * `filesystem` and "filesystem"

+ Modifies README to reflect usage of `null`.