[TEST] Mock `Configuration` with new helper.
Huzzah! All (entry) tests now mock the `Configuration` using the helper,
so now the default configuration is always standard, followed by any
changes specified in the decorator.

This should mean far more stable tests, and the `entry_mock` helper
method paves the way for unit-tests that actually test in units :-)
(see the disk tests for a nice example of this!)

* Also moves `distro` tests into the `entries` subfolder.