[CORE] Adds support for custom entries
This patch brings a first support for "custom entries" to Archey. It
mainly allows users to run custom commands through their configuration
file.

In order to give "power to the user" for most of situations:
* STDERR output **MAY** be silenced through `log_stderr` option
* command exit status code **MAY** be ignored through `check` option

Multi-lines results will be joined on the same line, unless `one_line`
option is disabled.

Users **SHOULD** only enable `shell` option if they have to.

> See #111 for rationale.