[CUSTOM][SECURITY] Adds "LPE" attacks mitigation
As new "custom" entries brings arbitrary command execution capabilities
in Archey code-base, we enforce some basic security checks to assert
loaded configuration files are not too "permissive".
The idea is to keep "W ^ X" enforced, so if another user _could_ write
to one of the configuration files, defined custom entries won't be
loaded.

Archey will consider a configuration file not too "permissive" if it
belongs to the running user (or the system administrator) **AND** if
group nor other users can write to it.